LayoutTests/storage/indexeddb/modern/opendatabase-after-storage-crash-expected.txt

Database connection closed by network process crash.
PASS window.indexedDB is the same object after crash
Successfully opened database with version 2 after network process crash.
PASS indexedDB.open() recovered after network process crash without page reload
PASS successfullyParsed is true

TEST COMPLETE
This test verifies when the network process crashes, indexedDB.open() recovers without a page reload.
